Added unregisteration route to authentication blueprint
