+++ api/templates/account.html
@@ -27,7 +27,7 @@
         
         <div class="action-block">
           <h2>Account Options</h2>
-          <a href="{{ url_for('unregister') }}" title="Delete your account" class="button delete-account">Delete Account</a>
+          <a href="{{ url_for('auth.unregister') }}" title="Delete your account" class="button delete-account">Delete Account</a>
         </div>
       
       </div>
